+++ src/app/app.routes.ts
@@ -4,12 +4,12 @@ export const routes: Routes = [
   {
     path: 'characters',
     title: 'Star Wars Characters',
-    loadComponent: () => import('./characters/characters-list.component').then(c => c.CharactersListComponent)
+    loadComponent: () => import('./characters-list/characters-list.component').then(c => c.CharactersListComponent)
   },
   {
     path: 'character-details/:id',
     title: 'Star Wars Character Details',
-    loadComponent: () => import('./characters/characters-list.component').then(c => c.CharactersListComponent)
+    loadComponent: () => import('./characters-list/characters-list.component').then(c => c.CharactersListComponent)
   },
   { path: '**', redirectTo: '/characters', pathMatch: 'full' },
 ];
