+++ src/app/characters-list/characters-list-item/characters-list-item.component.scss
@@ -1,5 +1,5 @@
 :host {
-  padding: 1rem;
+  padding: 0;
   border-radius: 1rem;
   margin-bottom: 1rem;
   box-shadow: 0.25rem 0.25rem 0.75rem rgb(0 0 0 / 0.1);
@@ -25,4 +25,7 @@ a {
   text-decoration: inherit;
   cursor: inherit;
   color: inherit;
+  width: 100%;
+  height: 100%;
+  padding: 1rem;
 }
