+++ src/app/characters-list/characters-list-item/characters-list-item.component.spec.ts
