+++ src/app/characters-list/characters-list.component.scss
