+++ src/app/characters-list/characters-pagination/characters-pagination.component.scss
